renamed temporary view frames
